Fixed database compilation and renamed paths in data processing scripts.

+++ scripts/stage2/extract.sh
@@ -1,20 +1,24 @@
 #!/bin/bash
-
+HELPTEXT="Usage: ./extract.sh <path to stage1 results> <path to directory containing ZIP files>"
+if [ "$#" -ne 2 ]; then
+    echo $HELPTEXT
+    exit 1
+fi
+RESULT_DIR="$1/stage2"
 echo "Preparing directories..."
-rm -rf ./dataset/stage2
+rm -rf $RESULT_DIR
 rm -rf ./temp
 
-RESULT_DIR=./dataset/stage2
 mkdir -p "$RESULT_DIR"
 mkdir "$RESULT_DIR/info"
 mkdir "$RESULT_DIR/logs"
 touch "$RESULT_DIR/visited.csv"
-touch "$RESULT_DIR/failed_download.csv"
+touch "$RESULT_DIR/status_download.csv"
 touch "$RESULT_DIR/status_miri_comp.csv"
 touch "$RESULT_DIR/status_rustc_comp.csv"
 touch "$RESULT_DIR/tests.csv"
 RESULT_DIR=./dataset/stage2
-for file in "$1"/*.zip; do
+for file in "$2"/*.zip; do
     unzip -q "$file"
     ROOT="./dataset/stage2"
     echo "$file"
@@ -28,7 +32,7 @@ for file in "$1"/*.zip; do
         done < "$csv_file"
     done
     cat $ROOT/visited.csv >> $RESULT_DIR/visited.csv
-    cat $ROOT/failed_download.csv >> $RESULT_DIR/failed_download.csv
+    cat $ROOT/status_download.csv >> $RESULT_DIR/status_download.csv
     cat $ROOT/status_miri_comp.csv >> $RESULT_DIR/status_miri_comp.csv
     cat $ROOT/status_rustc_comp.csv >> $RESULT_DIR/status_rustc_comp.csv
     cp -r $ROOT/logs/* $RESULT_DIR/logs
